% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_collection.R
\name{book_ids}
\alias{book_ids}
\title{List the books in a scan collection}
\usage{
book_ids(x)
}
\arguments{
\item{x}{A \code{\link[=scan_collection]{scan_collection()}}.}
}
\value{
A character vector of unique \code{book_id} values.
}
\description{
List the books in a scan collection
}
\examples{
df <- data.frame(
  book_id = c("book_1", "book_2"),
  chip = "10YR 8/1",
  L = 81.3,
  a = 1.3,
  b = 6.6
)
book_ids(scan_collection(df, sensor = "nix"))
}
