% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chipfinish}
\alias{chipfinish}
\title{Chip Finish}
\format{
a data.frame with 102 rows and 2 variables
\describe{
\item{chip}{The Munsell notation of the color chip}
\item{finish}{The percieved colorchip finish}
}
}
\source{
TBD
}
\usage{
chipfinish
}
\description{
This data.frame contains the percepted "finish" of the chips scanned
from the reference munsell soil color books. The manufacturer of these
books can finish the chip in one of \code{matte}, \code{semigloss}, or \code{gloss}.
Due to limitations in most colorimeters, it is difficult to measure glossy
chips accurately and repeatably. Further, it would seem these chips may
degrade the quickest. For these reasons, the finish data is included to
help users make appropriate decisions regarding the quality of their color books.
Users may provide their own glossiness data if their books finishes differ from
this dataset.
}
\keyword{datasets}
