% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colordata}
\alias{colordata}
\title{Color Data}
\format{
A data.frame with 412 rows and 5 variables
\describe{
\item{sensor}{The color sensor name}
\item{chip}{The Munsell notation of the color chip}
\item{L}{CIELab L value}
\item{a}{CIELab a value}
\item{b}{CIELab b value}
}
}
\source{
TBD
}
\usage{
colordata
}
\description{
This data.frame contains the aggregated color scan measurements of
three brand-new Munsell Soil Color books (2.5Y, 10YR, and 7.5YR only for now).
This dataset is used for comparisons with user scans, and is treated as the 'Truth'
when judging user books.
}
\keyword{datasets}
