% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_collection.R
\name{is_scan_collection}
\alias{is_scan_collection}
\title{Test whether an object is a \code{ScanCollection}}
\usage{
is_scan_collection(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Test whether an object is a \code{ScanCollection}
}
\examples{
is_scan_collection(1:10)
}
