% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_data.R
\docType{package}
\name{munqc-package}
\alias{munqc}
\alias{munqc-package}
\title{munqc: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
