% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_collection.R
\name{subset_ids}
\alias{subset_ids}
\title{Select books from a scan collection}
\usage{
subset_ids(x, book_id, negate = FALSE)
}
\arguments{
\item{x}{A \code{\link[=scan_collection]{scan_collection()}}.}

\item{book_id}{A \code{book_id} or a vector of them.}

\item{negate}{If \code{TRUE}, return every book except those named.}
}
\value{
A \code{ScanCollection}
}
\description{
Subsets a \code{\link[=scan_collection]{scan_collection()}} to one or more books, including results,
so this can be used before or after \code{\link[=compute_error]{compute_error()}}.
}
\examples{
d <- rbind(
  data.frame(book_id = "A", chip = "10YR 8/1", L = 81.3, a = 1.3, b = 6.6),
  data.frame(book_id = "B", chip = "10YR 8/1", L = 79.0, a = 1.1, b = 6.2),
  data.frame(book_id = "C", chip = "10YR 8/1", L = 80.1, a = 1.4, b = 6.9)
)
sc <- scan_collection(d, sensor = "nix")

books(sc, "A")
books(sc, c("A", "B"))
books(sc, "A", negate = TRUE)

}
\seealso{
\code{\link[=book_ids]{book_ids()}} to see books available.
}
