% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ScanCollection}
\alias{plot.ScanCollection}
\title{Plot a scored scan collection}
\usage{
\method{plot}{ScanCollection}(
  x,
  type = c("histogram", "page", "page_space"),
  binwidth = 0.25,
  scale = c("log", "linear"),
  facet_books = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{ScanCollection} that has been through \code{\link[=compute_error]{compute_error()}}.}

\item{type}{One of \code{"histogram"}, \code{"page"}, \code{"page_space"}.
Defaults to "\code{histogram}".}

\item{binwidth}{Histogram bin width. Defaults to \code{0.25}.}

\item{scale}{The scale of the dE axis, one of \code{"log"}, \code{"linear"}. Defaults
to \code{"log"}.}

\item{facet_books}{Facet by \code{book_id} when more than one book is present.
Defaults to \code{TRUE}.}

\item{...}{Ignored.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a scored scan collection
}
\examples{
\dontrun{
sc <- compute_error(scan_collection(my_scan_data, sensor = "nix"))
plot(sc)
plot(sc, binwidth = 0.3)
plot(sc, type = "page")
plot(sc, type = "page_space", facet_books = TRUE)
}

}
